Move knn files and utils
